[Mod]将BarGenerator和ArrayManager移动到vtUtility.py中，Close #1084
